- Move Columns class to a different file - Fix some issues in dependencies
