+++ package.json
@@ -38,7 +38,7 @@
     "express-winston": "^3.1.0",
     "fp-ts": "^1.9.0",
     "helmet": "^3.18.0",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.15",
     "ng-recaptcha": "^4.0.0-beta.1",
     "ng2-cookies": "^1.0.2",
     "ngx-bootstrap": "^3.1.4",
